move unneccessarily nested file
