+++ seqr/management/commands/check_for_new_samples_from_pipeline.py
@@ -13,7 +13,7 @@ from reference_data.models import GENOME_VERSION_LOOKUP
 from seqr.models import Family, Sample, Project, Individual, SavedVariant
 from seqr.utils.communication_utils import safe_post_to_slack, send_project_email
 from seqr.utils.file_utils import file_iter, list_files, is_google_bucket_file_path
-from seqr.utils.search.add_data_utils import notify_search_data_loaded, update_airtable_loading_tracking_status
+from seqr.utils.add_data_utils import notify_search_data_loaded, update_airtable_loading_tracking_status
 from seqr.views.utils.airtable_utils import AirtableSession, LOADABLE_PDO_STATUSES, AVAILABLE_PDO_STATUS
 from seqr.views.utils.export_utils import write_multiple_files
 from seqr.views.utils.permissions_utils import is_internal_anvil_project, project_has_anvil
